DataSafePlus app version with auto sign in
You will have to create a new vb.net Windows Desktop app using .NET 8 or higher. Controls are programmatically added in frmMain. Place the ds*.* files in the app folder and add them to the app tree. Also, use your own icons.
This version of the app attempts to log in to the website automatically and should work if the web site contains the username and password boxes.
Due to the app using web automation , you will have to generate a local certificate on your PC to satisfy the new Windows Safe App feature.