DataSafePlus app with Manual login
Create your own Windows desktop vb.net app using .NET 8 or higher. Add the ds*.* files to the app. This version does not use web automation and therefore does not need to be signed.  The user can mouse-middle-click on the Username and Password columns cells to automatically paste the contents to the web page current active control element. Clicking on the Web Page cell will attempt to open a web browser to the page location in the cell.